this brings our query in line with the previous version's flight query. we now query one flight segment, which results in one flight.

### original acmeair
```
Running 30s test @ http://localhost:9080/rest/api/flights/queryflights
10 connections

Stat         Avg    Stdev   Max    
Latency (ms) 16.18  7.95    212.33 
Req/Sec      599.37 46.6    677    
Bytes/Sec    472 kB 38.1 kB 533 kB 

18k requests in 30s, 14.2 MB read
null { title: 'acmeair-flights-old-code',
  url: 'http://localhost:9080/rest/api/flights/queryflights',
  socketPath: undefined,
  requests: 
   { average: 599.37,
     mean: 599.37,
     stddev: 46.6,
     min: 488,
     max: 677,
     total: 17981,
     sent: 17991 },
  latency: 
   { average: 16.18,
     mean: 16.18,
     stddev: 7.95,
     min: 3,
     max: 212.334212,
     p50: 15,
     p75: 19,
     p90: 24,
     p99: 38,
     p999: 61,
     p9999: 207,
     p99999: 212 },
  throughput: 
   { average: 471586.14,
     mean: 471586.14,
     stddev: 38054.37,
     min: 384056,
     max: 532799,
     total: 14151047 },
  errors: 0,
  timeouts: 0,
  duration: 30,
  start: 2018-06-27T11:40:16.140Z,
  finish: 2018-06-27T11:40:46.166Z,
  connections: 10,
  pipelining: 1,
  non2xx: 0,
  '1xx': 0,
  '2xx': 17981,
  '3xx': 0,
  '4xx': 0,
  '5xx': 0 }
```

### new version results
```
Running 30s test @ http://localhost:9080/rest/api/flights
10 connections

Stat         Avg    Stdev   Max     
Latency (ms) 14.7   23.48   1012.77 
Req/Sec      657.67 124.01  747     
Bytes/Sec    104 kB 19.7 kB 118 kB  

20k requests in 30s, 3.12 MB read
null { title: 'acmeair-flights',
  url: 'http://localhost:9080/rest/api/flights',
  socketPath: undefined,
  requests: 
   { average: 657.67,
     mean: 657.67,
     stddev: 124.01,
     min: 21,
     max: 747,
     total: 19730,
     sent: 19740 },
  latency: 
   { average: 14.7,
     mean: 14.7,
     stddev: 23.48,
     min: 2,
     max: 1012.772756,
     p50: 13,
     p75: 18,
     p90: 23,
     p99: 39,
     p999: 61,
     p9999: 1009,
     p99999: 1012 },
  throughput: 
   { average: 103805.87,
     mean: 103805.87,
     stddev: 19693.59,
     min: 3318,
     max: 118026,
     total: 3117340 },
  errors: 0,
  timeouts: 0,
  duration: 30,
  start: 2018-06-27T13:26:27.561Z,
  finish: 2018-06-27T13:26:57.592Z,
  connections: 10,
  pipelining: 1,
  non2xx: 0,
  '1xx': 0,
  '2xx': 19730,
  '3xx': 0,
  '4xx': 0,
  '5xx': 0 }
```

related: #3