@robbdempsey 2018-06-27
`clinic upload 21368.clinic-doctor`

![screen shot 2018-06-27 at 8 51 55 am](https://user-images.githubusercontent.com/2875104/41978735-0a883d70-79e8-11e8-8156-f1614b27284a.png)